refactor: separate decorator tests into files
