+++ package.json
@@ -71,6 +71,7 @@
     "gulp-rename": "^1.2.2",
     "gulp-uglify": "^3.0.0",
     "gzip-size-cli": "^3.0.0",
+    "husky": "^1.1.4",
     "jquery": "^3.0.0",
     "karma": "^3.0.0",
     "karma-browserify": "^5.2.0",
@@ -78,7 +79,7 @@
     "karma-coverage": "^1.1.0",
     "karma-firefox-launcher": "^1.0.0",
     "karma-mocha": "^1.0.1",
-    "lint-staged": "^8.0.0",
+    "lint-staged": "^8.0.5",
     "merge-stream": "^1.0.1",
     "mocha": "^5.0.0",
     "npm-run-all": "^4.1.2",
@@ -120,7 +121,7 @@
   },
   "lint-staged": {
     "*.js": [
-      "prettier --write --no-semi --single-quote --print-width 1000",
+      "prettier --write --no-semi --single-quote",
       "standard --fix",
       "git add"
     ]
