--- src/decorators/__tests__/decorators.js
@@ -1,155 +1,0 @@
-// @flow
-
-import assert from 'assert'
-import { div } from 'dom-gen'
-import { def, make, emits, component, wired } from '../../'
-import { clearComponents } from '../../__tests__/helper'
-import {
-  callClassDecorator,
-  callMethodDecorator
-} from './helper'
-
-describe('@emits(event)', () => {
-  afterEach(() => clearComponents())
-
-  it('throws when the empty event is given', () => {
-    class Component {
-      emitter () {}
-    }
-
-    def('component', Component)
-
-    assert.throws(() => {
-      callMethodDecorator(emits(undefined), Component, 'emitter')
-    }, /Unable to emits an empty event: constructor=Component key=emitter/)
-  })
-
-  it('makes the method emit the event with the returned value', done => {
-    class Component {
-      foo () {
-        return 321
-      }
-    }
-    def('component', Component)
-    callMethodDecorator(emits('event-foo'), Component, 'foo')
-
-    make(
-      'component',
-      div().on('event-foo', e => {
-        assert(e.detail === 321)
-
-        done()
-      })[0]
-    ).foo()
-  })
-
-  it('makes the method emit the event with the resolved value after the promise resolved', done => {
-    let promiseResolved = false
-
-    class Component {
-      foo () {
-        return new Promise(resolve => {
-          setTimeout(() => {
-            promiseResolved = true
-            resolve(123)
-          }, 100)
-        })
-      }
-    }
-    def('component', Component)
-    callMethodDecorator(emits('event-foo'), Component, 'foo')
-
-    make(
-      'component',
-      div().on('event-foo', e => {
-        assert(promiseResolved)
-        assert(e.detail === 123)
-
-        done()
-      })[0]
-    ).foo()
-  })
-})
-
-describe('@component(name)', () => {
-  afterEach(() => clearComponents())
-
-  it('works as a class decorator and registers the class as a class component of the given name', () => {
-    class Foo {
-      el: HTMLElement
-
-      __mount__ () {
-        this.el.setAttribute('this-is', 'decorated-component')
-      }
-    }
-
-    callClassDecorator(component('decorated-component'), Foo)
-
-    const el = div()[0]
-
-    make('decorated-component', el)
-
-    assert(el.getAttribute('this-is') === 'decorated-component')
-  })
-})
-
-describe('@wired.component(name, selector)', () => {
-  afterEach(() => clearComponents())
-
-  it('replaces the getter of the decorated descriptor, and it returns the instance of class-component inside the element', () => {
-    class Cls0 {
-      @wired.component('wire-test1-1', '.foo')
-      test = {}
-    }
-    class Cls1 {}
-    def('wire-test1', Cls0)
-    def('wire-test1-1', Cls1)
-
-    const el = div().append(
-      make('wire-test1-1', div().addClass('foo')[0]).el
-    )[0]
-
-    const wireTest1 = make('wire-test1', el)
-
-    assert(wireTest1.test instanceof Cls1)
-  })
-})
-
-describe('@wired(selector)', () => {
-  afterEach(() => clearComponents())
-
-  it('wires the element in the component', () => {
-    class Component {
-      @wired('.elm')
-      elm = {}
-    }
-
-    def('wire-el-test', Component)
-
-    const el = div().append(div().addClass('elm'))[0]
-
-    const component = make('wire-el-test', el)
-    assert(component.elm instanceof HTMLElement)
-    assert(component.elm === el.firstChild)
-  })
-})
-
-describe('@wired.all(selector)', () => {
-  afterEach(() => clearComponents())
-
-  it('wires the all elements in the component', () => {
-    class Component {
-      @wired.all('.elm')
-      elms = []
-    }
-
-    def('comp', Component)
-
-    const el = div().append(div().addClass('elm'), div().addClass('elm'))[0]
-
-    const component = make('comp', el)
-    assert(component.elms.length === 2)
-    assert(component.elms[0] === el.firstChild)
-    assert(component.elms[1] === el.lastChild)
-  })
-})
